refactor: get, post, util 파일별로 분리
