+++ main.ts
@@ -6,42 +6,14 @@ import { Prisma, PrismaClient } from "@prisma/client"
 import { serve } from "@hono/node-server"
 import { z } from "zod"
 
-import { ApplicationSchema, PositionSchema, TechSchema } from "./prisma/generated/zod/index.ts"
+import { ApplicationSchema, PositionSchema } from "./prisma/generated/zod/index.ts"
 
 import { NumericSchema } from "./data"
 import { createHttpTerminator } from "http-terminator"
 import { Server } from "node:http"
 import { showUniqueRoutes } from "./unique_routes.ts"
 import { isMain } from "./is_main.ts"
-import { getSinglePosition } from "./service.ts"
-import type { NonEmptyObject } from "type-fest"
-
-export const select = {
-	id: true,
-	title: true,
-	reward: true,
-	techStack: true,
-	company: true,
-	region: true,
-} satisfies Prisma.PositionSelect
-
-export const fuzzySearch = (search?: string) =>
-	search
-		? ({
-			OR: [
-				{ company: { name: { contains: search } } },
-				{ title: { contains: search } },
-				{ description: { contains: search } },
-				{ techStack: { some: { name: { contains: search } } } },
-			],
-		} satisfies Prisma.PositionWhereInput)
-		: undefined
-
-export const nonEmpty = <Schema extends z.AnyZodObject>(schema: Schema) =>
-	schema.partial().refine(
-		(v): v is NonEmptyObject<Partial<Schema["_type"]>> => Object.keys(v).length !== 0,
-		{ message: `최소 ${Object.keys(schema.shape)}중 하나의 필드가 필요합니다.` },
-	)
+import { positions } from "./positions/mod.ts"
 
 type Option = { prisma: PrismaClient; app: Hono }
 export const createApp = ({ prisma, app }: Option) =>
@@ -53,23 +25,7 @@ export const createApp = ({ prisma, app }: Option) =>
 
 			return c.jsonT(result)
 		})
-		.post(
-			"/positions",
-			zValidator(
-				"json",
-				PositionSchema
-					.omit({ id: true })
-					.extend({ techStack: z.array(nonEmpty(TechSchema)) }),
-			),
-			async (c) => {
-				const { techStack, ...rest } = c.req.valid("json")
-				const result = await prisma.position.create({
-					data: { ...rest, techStack: { connect: techStack } },
-				})
-
-				return c.jsonT(result)
-			},
-		)
+        .route("/positions", positions(prisma))
 		.patch(
 			"/positions/:id",
 			zValidator("param", z.object({ id: NumericSchema })),
@@ -92,27 +48,6 @@ export const createApp = ({ prisma, app }: Option) =>
 				const result = await prisma.position.delete({ where: { id } })
 
 				return c.jsonT(result)
-			},
-		)
-		.get(
-			"/positions",
-			zValidator("query", z.object({ search: z.string().min(2).optional() })),
-			async (c) => {
-				const { search } = c.req.valid("query")
-
-				const result = await prisma.position.findMany({ select, where: fuzzySearch(search) })
-
-				return c.jsonT(result)
-			},
-		)
-		.get(
-			"/positions/:id",
-			zValidator("param", z.object({ id: NumericSchema })),
-			async (c) => {
-				const id = c.req.valid("param").id
-
-				const { result, otherPositions } = await getSinglePosition(prisma, id)
-				return result ? c.jsonT({ ...result, otherPositions }) : c.notFound()
 			},
 		)
 		.onError((error, c) => {
