+++ package.json
@@ -8,7 +8,7 @@
 		"db:init": "npm exec prisma migrate dev -- --name dev",
 		"prod": "node --no-warnings=ExperimentalWarning --loader @swc-node/register/esm main.ts",
 		"dev": "node --no-warnings=ExperimentalWarning --loader @swc-node/register/esm --watch main.ts",
-		"test": "node --no-warnings=ExperimentalWarning --loader @swc-node/register/esm --test get_test.ts post_test.ts"
+		"test": "node --no-warnings=ExperimentalWarning --loader @swc-node/register/esm --test **/*_test.ts"
 	},
 	"prisma": {
 		"seed": "node --no-warnings=ExperimentalWarning --loader @swc-node/register/esm seed.ts"
