+++ positions/get_test.ts
@@ -1,9 +1,9 @@
 import test from "node:test"
 import assert from "node:assert/strict"
-import { createApp } from "./main.ts"
-import { tmpPrismaClient } from "./tmp_prisma_client.ts"
+import { createApp } from "../main.ts"
+import { tmpPrismaClient } from "../tmp_prisma_client.ts"
 import { Hono } from "hono/quick"
-import { serverUrl, tempDbFileUrl } from "./test_url.ts"
+import { serverUrl, tempDbFileUrl } from "../test_url.ts"
 import { testClient } from "hono/testing"
 
 test("GET /positions", { concurrency: true }, async (t) => {
